requirements.txt
httpx
html2text
playwright
anthropic
openai
google-genai
gspread
pydantic
python-dotenv
click
